@@ -2,6 +2,7 @@ package step
 
 import (
 	"bytes"
+	"context"
 	"strings"
 	"testing"
 )
@@ -11,13 +12,13 @@ func TestJsonStep_Invoke(t *testing.T) {
 	input := `{"foo":"bar"}`
 	config := `{"template":"{\"baz\":\"{{ .foo }}\"}"}`
 
-	step,err := UnmarshalJsonStep(strings.NewReader(config))
+	step, err := UnmarshalJsonStep(strings.NewReader(config))
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	buf := bytes.Buffer{}
-	err = step.Invoke(strings.NewReader(input),&buf)
+	err = step.Invoke(context.Background(), strings.NewReader(input), &buf)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -25,7 +26,7 @@ func TestJsonStep_Invoke(t *testing.T) {
 	got := buf.String()
 	want := `{"baz":"bar"}`
 	if want != got {
-		t.Fatalf("want: %s\ngot: %s\n",want,got)
+		t.Fatalf("want: %s\ngot: %s\n", want, got)
 	}
 }
 
@@ -34,13 +35,13 @@ func TestJsonStep_Invoke_NewJSON(t *testing.T) {
 	input := `{"foo":"bar"}`
 	tmpl := "{\"baz\":\"{{ .foo }}\"}"
 
-	step,err := NewJSON(tmpl)
+	step, err := NewJSON(tmpl)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	buf := bytes.Buffer{}
-	err = step.Invoke(strings.NewReader(input),&buf)
+	err = step.Invoke(context.Background(), strings.NewReader(input), &buf)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -48,6 +49,6 @@ func TestJsonStep_Invoke_NewJSON(t *testing.T) {
 	got := buf.String()
 	want := `{"baz":"bar"}`
 	if want != got {
-		t.Fatalf("want: %s\ngot: %s\n",want,got)
+		t.Fatalf("want: %s\ngot: %s\n", want, got)
 	}
 }
