@@ -2,6 +2,7 @@ package step
 
 import (
 	"bytes"
+	"context"
 	"testing"
 )
 
@@ -9,11 +10,11 @@ func TestNoOpStep_Invoke(t *testing.T) {
 	step := NoOpStep{}
 	in := []byte("noop")
 	out := bytes.Buffer{}
-	err := step.Invoke(bytes.NewReader(in),&out)
+	err := step.Invoke(context.Background(), bytes.NewReader(in), &out)
 	if err != nil {
 		t.Fatal(err)
 	}
 	if out.String() != "noop" {
-		t.Fatalf("want: noop, got: %s\n",out.String())
+		t.Fatalf("want: noop, got: %s\n", out.String())
 	}
 }
